move payment types to pallet
